add trait for slack escape
